refactor(rpc): dedupe auth test helpers
