+++ crates/rpc/tests/it/ws.rs
@@ -1,26 +1,15 @@
-use std::{
-    collections::HashMap,
-    sync::{Arc, OnceLock},
-};
+use std::{collections::HashMap, sync::Arc};
 
-use alloy_primitives::{Address, B256, Bytes};
+use alloy_primitives::Address;
 use alloy_signer::SignerSync;
 use alloy_signer_local::PrivateKeySigner;
-use base64::Engine as _;
 use futures::{SinkExt, StreamExt};
-use p256::{
-    EncodedPoint,
-    ecdsa::{SigningKey as P256SigningKey, signature::hazmat::PrehashSigner},
-};
+use p256::ecdsa::SigningKey as P256SigningKey;
 use parking_lot::Mutex;
 use rand::thread_rng;
 use serde_json::Value;
-use sha2::{Digest, Sha256};
 use tempo_contracts::precompiles::account_keychain::IAccountKeychain::{
     KeyInfo, SignatureType as KeyInfoSignatureType,
-};
-use tempo_primitives::transaction::tt_signature::{
-    KeychainSignature, PrimitiveSignature, TempoSignature, WebAuthnSignature, normalize_p256_s,
 };
 use tokio_tungstenite::{connect_async, tungstenite};
 use zone_rpc::{
@@ -31,10 +20,19 @@ use zone_rpc::{
     types::{BoxEyreFut, BoxFut, JsonRpcError},
 };
 
+#[path = "../../test-utils/auth_tokens.rs"]
+mod auth_tokens;
+
+use auth_tokens::{
+    build_token_with_signature, now_secs, sign_keychain_signature, sign_p256_signature,
+    sign_webauthn_signature,
+};
+
 // ---------------------------------------------------------------------------
 // Mock API
 // ---------------------------------------------------------------------------
 
+#[derive(Default)]
 struct MockZoneRpcApi {
     key_infos: Mutex<HashMap<(Address, Address), KeyInfo>>,
     key_lookup_error: Option<&'static str>,
@@ -154,11 +152,7 @@ impl TestContext {
     }
 
     fn build_token(&self) -> String {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap()
-            .as_secs();
-
+        let now = now_secs();
         let (fields, digest) = build_token_fields(ZONE_ID, CHAIN_ID, PORTAL, now, now + 600);
         let sig = self.signer.sign_hash_sync(&digest).expect("signing failed");
 
@@ -174,111 +168,6 @@ impl TestContext {
     fn ws_url(&self) -> String {
         format!("ws://{}", self.addr)
     }
-}
-
-fn default_api() -> MockZoneRpcApi {
-    MockZoneRpcApi {
-        key_infos: Mutex::new(HashMap::new()),
-        key_lookup_error: None,
-    }
-}
-
-fn p256_public_key(signing_key: &P256SigningKey) -> (B256, B256) {
-    let encoded = EncodedPoint::from(signing_key.verifying_key());
-    let pub_key_x = B256::from_slice(encoded.x().expect("x coordinate present"));
-    let pub_key_y = B256::from_slice(encoded.y().expect("y coordinate present"));
-    (pub_key_x, pub_key_y)
-}
-
-fn sign_p256_token(digest: B256, signing_key: &P256SigningKey) -> eyre::Result<TempoSignature> {
-    let pre_hashed = Sha256::digest(digest);
-    let signature: p256::ecdsa::Signature = signing_key.sign_prehash(&pre_hashed)?;
-    let sig_bytes = signature.to_bytes();
-    let (pub_key_x, pub_key_y) = p256_public_key(signing_key);
-
-    Ok(TempoSignature::Primitive(PrimitiveSignature::P256(
-        tempo_primitives::transaction::tt_signature::P256SignatureWithPreHash {
-            r: B256::from_slice(&sig_bytes[0..32]),
-            s: normalize_p256_s(&sig_bytes[32..64]),
-            pub_key_x,
-            pub_key_y,
-            pre_hash: true,
-        },
-    )))
-}
-
-fn sign_webauthn_token(
-    signing_key: &P256SigningKey,
-    challenge: B256,
-) -> eyre::Result<TempoSignature> {
-    let mut authenticator_data = vec![0u8; 37];
-    authenticator_data[0..32].copy_from_slice(&[0xAA; 32]);
-    authenticator_data[32] = 0x01;
-    let challenge_b64url = base64::engine::general_purpose::URL_SAFE_NO_PAD.encode(challenge);
-    let client_data_json = format!(
-        r#"{{"type":"webauthn.get","challenge":"{challenge_b64url}","origin":"https://example.com","crossOrigin":false}}"#
-    );
-    let client_data_hash = Sha256::digest(client_data_json.as_bytes());
-    let mut final_hasher = Sha256::new();
-    final_hasher.update(&authenticator_data);
-    final_hasher.update(client_data_hash);
-    let message_hash = final_hasher.finalize();
-
-    let signature: p256::ecdsa::Signature = signing_key.sign_prehash(&message_hash)?;
-    let sig_bytes = signature.to_bytes();
-    let (pub_key_x, pub_key_y) = p256_public_key(signing_key);
-    let mut webauthn_data = authenticator_data;
-    webauthn_data.extend_from_slice(client_data_json.as_bytes());
-
-    Ok(TempoSignature::Primitive(PrimitiveSignature::WebAuthn(
-        WebAuthnSignature {
-            webauthn_data: Bytes::from(webauthn_data),
-            r: B256::from_slice(&sig_bytes[0..32]),
-            s: normalize_p256_s(&sig_bytes[32..64]),
-            pub_key_x,
-            pub_key_y,
-        },
-    )))
-}
-
-fn sign_keychain_token(
-    digest: B256,
-    access_signer: &P256SigningKey,
-    root_account: Address,
-    version: u8,
-) -> eyre::Result<(TempoSignature, Address)> {
-    let signing_hash = match version {
-        0x03 => digest,
-        0x04 => KeychainSignature::signing_hash(digest, root_account),
-        _ => eyre::bail!("unsupported keychain version"),
-    };
-    let primitive = match sign_p256_token(signing_hash, access_signer)? {
-        TempoSignature::Primitive(primitive) => primitive,
-        TempoSignature::Keychain(_) => unreachable!("primitive signature expected"),
-    };
-    let signature = if version == 0x03 {
-        TempoSignature::Keychain(KeychainSignature::new_v1(root_account, primitive))
-    } else {
-        TempoSignature::Keychain(KeychainSignature::new(root_account, primitive))
-    };
-    let key_id = signature
-        .recover_signer(&digest)
-        .expect("keychain signer recovery should succeed");
-    let access_key_id = match &signature {
-        TempoSignature::Keychain(keychain) => keychain
-            .key_id(&digest)
-            .expect("inner key recovery should succeed"),
-        TempoSignature::Primitive(_) => unreachable!("keychain signature expected"),
-    };
-    assert_eq!(key_id, root_account);
-    Ok((signature, access_key_id))
-}
-
-fn build_token_with_signature(signature: TempoSignature, fields: &[u8]) -> String {
-    let mut blob = Vec::with_capacity(signature.encoded_length() + fields.len());
-    blob.extend_from_slice(signature.to_bytes().as_ref());
-    blob.extend_from_slice(fields);
-    alloy_primitives::hex::encode(&blob)
 }
 
 /// Build a JSON-RPC request string.
@@ -329,19 +218,13 @@ fn parse_response(msg: tungstenite::Message) -> Value {
     }
 }
 
-fn test_lock() -> &'static tokio::sync::Mutex<()> {
-    static LOCK: OnceLock<tokio::sync::Mutex<()>> = OnceLock::new();
-    LOCK.get_or_init(|| tokio::sync::Mutex::new(()))
-}
-
 // ---------------------------------------------------------------------------
 // Tests
 // ---------------------------------------------------------------------------
 
 #[tokio::test]
 async fn ws_roundtrip_with_header_auth() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let mut ws = connect_with_header(&ctx).await;
 
     ws.send(tungstenite::Message::Text(
@@ -357,8 +240,7 @@ async fn ws_roundtrip_with_header_auth() {
 
 #[tokio::test]
 async fn ws_roundtrip_with_query_auth() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let token = ctx.build_token();
     let url = format!("{}/?token={token}", ctx.ws_url());
 
@@ -377,8 +259,7 @@ async fn ws_roundtrip_with_query_auth() {
 
 #[tokio::test]
 async fn ws_reject_no_auth() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let result = connect_async(ctx.ws_url()).await;
     // Server should reject the upgrade — tungstenite surfaces this as an error
     // with the HTTP 401 status.
@@ -391,8 +272,7 @@ async fn ws_reject_no_auth() {
 
 #[tokio::test]
 async fn ws_reject_invalid_token() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let req = tungstenite::http::Request::builder()
         .uri(ctx.ws_url())
         .header("x-authorization-token", "deadbeef")
@@ -418,8 +298,7 @@ async fn ws_reject_invalid_token() {
 
 #[tokio::test]
 async fn ws_multiple_requests() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let mut ws = connect_with_header(&ctx).await;
 
     for i in 1..=5 {
@@ -436,8 +315,7 @@ async fn ws_multiple_requests() {
 
 #[tokio::test]
 async fn ws_batch_request() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let mut ws = connect_with_header(&ctx).await;
 
     let batch = serde_json::json!([
@@ -459,8 +337,7 @@ async fn ws_batch_request() {
 
 #[tokio::test]
 async fn ws_invalid_json() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let mut ws = connect_with_header(&ctx).await;
 
     ws.send(tungstenite::Message::Text("{broken".into()))
@@ -473,8 +350,7 @@ async fn ws_invalid_json() {
 
 #[tokio::test]
 async fn ws_unknown_method() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let mut ws = connect_with_header(&ctx).await;
 
     ws.send(tungstenite::Message::Text(jsonrpc("eth_foobar", 1).into()))
@@ -488,8 +364,7 @@ async fn ws_unknown_method() {
 
 #[tokio::test]
 async fn ws_disabled_method() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let mut ws = connect_with_header(&ctx).await;
 
     ws.send(tungstenite::Message::Text(
@@ -505,8 +380,7 @@ async fn ws_disabled_method() {
 
 #[tokio::test]
 async fn ws_empty_batch() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let mut ws = connect_with_header(&ctx).await;
 
     ws.send(tungstenite::Message::Text("[]".into()))
@@ -519,16 +393,12 @@ async fn ws_empty_batch() {
 
 #[tokio::test]
 async fn ws_roundtrip_with_p256_auth() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let signing_key = P256SigningKey::random(&mut thread_rng());
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap()
-        .as_secs();
+    let now = now_secs();
     let (fields, digest) = build_token_fields(ZONE_ID, CHAIN_ID, PORTAL, now, now + 600);
     let token = build_token_with_signature(
-        sign_p256_token(digest, &signing_key).expect("p256 signing should succeed"),
+        sign_p256_signature(digest, &signing_key).expect("p256 signing should succeed"),
         &fields,
     );
     let mut ws = connect_with_token(&ctx.ws_url(), ctx.addr, &token)
@@ -548,16 +418,12 @@ async fn ws_roundtrip_with_p256_auth() {
 
 #[tokio::test]
 async fn ws_roundtrip_with_webauthn_auth() {
-    let _guard = test_lock().lock().await;
-    let ctx = TestContext::start(default_api()).await;
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
     let signing_key = P256SigningKey::random(&mut thread_rng());
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap()
-        .as_secs();
+    let now = now_secs();
     let (fields, digest) = build_token_fields(ZONE_ID, CHAIN_ID, PORTAL, now, now + 600);
     let token = build_token_with_signature(
-        sign_webauthn_token(&signing_key, digest).expect("webauthn signing should succeed"),
+        sign_webauthn_signature(&signing_key, digest).expect("webauthn signing should succeed"),
         &fields,
     );
     let mut ws = connect_with_token(&ctx.ws_url(), ctx.addr, &token)
@@ -577,16 +443,12 @@ async fn ws_roundtrip_with_webauthn_auth() {
 
 #[tokio::test]
 async fn ws_roundtrip_with_keychain_auth() {
-    let _guard = test_lock().lock().await;
     let root_account = Address::repeat_byte(0x55);
     let access_signer = P256SigningKey::random(&mut thread_rng());
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap()
-        .as_secs();
+    let now = now_secs();
     let (fields, digest) = build_token_fields(ZONE_ID, CHAIN_ID, PORTAL, now, now + 600);
-    let (signature, key_id) =
-        sign_keychain_token(digest, &access_signer, root_account, 0x04).unwrap();
+    let (signature, key_id) = sign_keychain_signature(digest, &access_signer, root_account, 0x04)
+        .expect("keychain signing should succeed");
     let ctx = TestContext::start(MockZoneRpcApi::with_key(
         root_account,
         key_id,
@@ -611,17 +473,13 @@ async fn ws_roundtrip_with_keychain_auth() {
 
 #[tokio::test]
 async fn ws_reject_unauthorized_keychain_token() {
-    let _guard = test_lock().lock().await;
     let root_account = Address::repeat_byte(0x44);
     let access_signer = P256SigningKey::random(&mut thread_rng());
-    let ctx = TestContext::start(default_api()).await;
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap()
-        .as_secs();
+    let ctx = TestContext::start(MockZoneRpcApi::default()).await;
+    let now = now_secs();
     let (fields, digest) = build_token_fields(ZONE_ID, CHAIN_ID, PORTAL, now, now + 600);
-    let (signature, _key_id) =
-        sign_keychain_token(digest, &access_signer, root_account, 0x04).unwrap();
+    let (signature, _key_id) = sign_keychain_signature(digest, &access_signer, root_account, 0x04)
+        .expect("keychain signing should succeed");
     let token = build_token_with_signature(signature, &fields);
 
     let err = connect_with_token(&ctx.ws_url(), ctx.addr, &token)
@@ -635,17 +493,13 @@ async fn ws_reject_unauthorized_keychain_token() {
 
 #[tokio::test]
 async fn ws_keychain_lookup_failure_returns_500() {
-    let _guard = test_lock().lock().await;
     let root_account = Address::repeat_byte(0x66);
     let access_signer = P256SigningKey::random(&mut thread_rng());
     let ctx = TestContext::start(MockZoneRpcApi::with_key_lookup_error("key lookup failed")).await;
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap()
-        .as_secs();
+    let now = now_secs();
     let (fields, digest) = build_token_fields(ZONE_ID, CHAIN_ID, PORTAL, now, now + 600);
-    let (signature, _key_id) =
-        sign_keychain_token(digest, &access_signer, root_account, 0x04).unwrap();
+    let (signature, _key_id) = sign_keychain_signature(digest, &access_signer, root_account, 0x04)
+        .expect("keychain signing should succeed");
     let token = build_token_with_signature(signature, &fields);
 
     let err = connect_with_token(&ctx.ws_url(), ctx.addr, &token)
